style(vite-app): rename Field file
