+++ packages/vite-app/src/components/class/index.ts
@@ -1,4 +1,4 @@
 export * from "./ClassView";
 export * from "./Header";
 export * from "./Body";
-export * from "./Fields";
+export * from "./Field";
